requirements.txt
pygame
click
ruamel.yaml
